migrate to hatchling and nox (#44)
* migrate to hatchling and update metadata

* add noxfile to run tests and linters; fix errors

* CI: update workflows to use nox

* README: document new nox dev workflow

* CI: use pip for antsibull{,-docs} workflows

* CI antsibull: add explicit dependency on wheel

We need to manually add wheel here for `setup.py bdist_wheel` that
antsibull-build calls. wheel is not added by default for venvs created by
`python -m venv`. This should've been explicitly specifed as a
dependency of `antsibull`. We've fixed this problem in more recent
releases by switching to `build`.

* replace broad `Exception`s with more specific ones