FakeVenvRunner: replace python with sys.executable (#65)
* FakeVenvRunner: replace python with sys.executable

This way you can run `venv.async_log_run()` and always get the correct
python. In VenvRunner, it be `{prefix}/bin/python`. Outside, it should
be `sys.executable`; the first python on $PATH may not be the one that
the program was executed with.

* venv: use single quotes consistently

* VenvRunner log_run: don't mutate `args`

We should avoid messing with mutable arguments. Also, this brings the
function signature inline with the `async_log_run` function that accepts
any Sequence instead of only mutable ones.